refactor(account): split chain.go into modular handler package
- Create new pkg/handler package with chain of responsibility pattern
- Split large chain.go file into focused, single-responsibility handlers:
  - types.go: Common types and interfaces
  - base.go: Base handler implementation
  - validation.go: Account validation handlers
  - money.go: Money creation and conversion handlers
  - operation.go: Domain operation handlers
  - persistence.go: Persistence handlers
  - builder.go: Chain builder
  - chain.go: Simplified chain interface
- Maintain backward compatibility in account service
- Add comprehensive documentation in README.md
- Improve code organization, testability, and maintainability
- Reduce file complexity from 494 lines to focused modules

This refactoring improves separation of concerns, makes handlers
independently testable, and provides a cleaner architecture for
future extensions.
